Fix issue with file type displaying incorrectly in aggregate works with multiple file types. Consolidates file type determination in one place.
