+++ static/js/vue-cdr-access/src/components/listDisplay.vue
@@ -34,6 +34,7 @@ Renders search results in a list view display format
 <script>
     import thumbnail from '@/components/full_record/thumbnail.vue';
     import displayUtils from '../mixins/displayUtils';
+    import fileUtils from '../mixins/fileUtils';
     import permissionUtils from '../mixins/permissionUtils';
     import { format } from 'date-fns';
 
@@ -42,7 +43,7 @@ Renders search results in a list view display format
 
         components: {thumbnail},
 
-        mixins: [displayUtils, permissionUtils],
+        mixins: [displayUtils, fileUtils, permissionUtils],
 
         props: {
             recordList: {
@@ -90,35 +91,6 @@ Renders search results in a list view display format
                 }
 
                 return '';
-            },
-
-            /**
-             * @param record
-             * @returns {string} File Type value for a work or file object. Will return the descriptive
-             *      form of the file type if available, and fall back to the mimetype if not
-             */
-            getFileType(record) {
-                let fileType = this.determineFileType(record.fileDesc);
-                if (!fileType) {
-                    fileType = this.determineFileType(record.fileType);
-                }
-                return fileType || '';
-            },
-
-            /**
-             * Determines which filetype should be shown
-             * For multiple filetypes it de-dupes the array and if only one value show that, otherwise show 'Various'
-             * @param fileTypes
-             * @returns {string|*|undefined}
-             */
-            determineFileType(fileTypes) {
-                if (fileTypes && fileTypes.length === 1) {
-                    return fileTypes[0];
-                } else if (fileTypes && fileTypes.length > 1) {
-                    return ([...new Set(fileTypes)].length === 1) ? fileTypes[0] : 'Various';
-                } else {
-                    return undefined;
-                }
             },
 
             showCollection(record) {
